+++ CHANGELOG.md
@@ -86,7 +86,7 @@ v4.6
 - Support using swig version 4.2 to generate Java & Python bindings. (#4028)
 - Added `ExpressionBasedPathForce`, which can be used to create non-linear path springs or 
   other path-based force elements dependent on a user-provided expression. (#4035)
-
+- Fixed bugs in `PolynomialPathFitter` when too few coordinate samples were provided. (#4039)
 
 
 v4.5.1